pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "covision"
version = "0.1.0"
description = "Computer vision module for AI companions - presence detection, face recognition, gaze tracking"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Al Numan", email = "numan@example.com" }
]
keywords = [
    "computer-vision",
    "face-recognition",
    "presence-detection",
    "gaze-tracking",
    "ai-companion",
    "yolo",
    "mediapipe",
    "insightface"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Recognition",
]

dependencies = [
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "ultralytics>=8.3.0",
    "insightface>=0.7.3",
    "mediapipe>=0.10.0",
    "onnxruntime>=1.16.0",
    "pyyaml>=6.0",
    "pillow>=10.0.0",
]

[project.optional-dependencies]
gpu = [
    "onnxruntime-gpu>=1.16.0",
]
vlm = [
    "transformers>=4.40.0",
    "torch>=2.0.0",
    "accelerate>=0.25.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "black>=23.0.0",
]
all = [
    "covision[gpu,vlm,dev]",
]

[project.urls]
Homepage = "https://github.com/imnuman/co-vision"
Repository = "https://github.com/imnuman/co-vision"
Documentation = "https://github.com/imnuman/co-vision#readme"

[project.scripts]
covision = "covision.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/covision"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ["py311"]
